Move 'HttpClientTypeNotSupported' exception
