+++ src/Contracts/HTTP/HttpClientFactory.php
@@ -3,6 +3,7 @@
 namespace ArtARTs36\MergeRequestLinter\Contracts\HTTP;
 
 use ArtARTs36\MergeRequestLinter\Configuration\HttpClientConfig;
+use ArtARTs36\MergeRequestLinter\Exception\HttpClientTypeNotSupported;
 
 /**
  * Interface for HTTP Client Factory.
@@ -11,6 +12,7 @@ interface HttpClientFactory
 {
     /**
      * Create HTTP Client.
+     * @throws HttpClientTypeNotSupported
      */
     public function create(HttpClientConfig $config): Client;
 }
