+++ src/Support/Http/ClientFactory.php
@@ -7,7 +7,6 @@ use ArtARTs36\MergeRequestLinter\Contracts\HTTP\Client;
 use ArtARTs36\MergeRequestLinter\Contracts\HTTP\HttpClientFactory;
 use ArtARTs36\MergeRequestLinter\Contracts\Report\MetricManager;
 use ArtARTs36\MergeRequestLinter\Exception\HttpClientTypeNotSupported;
-use ArtARTs36\MergeRequestLinter\Tests\Mocks\NullClient;
 use GuzzleHttp\Client as GuzzleClient;
 
 class ClientFactory implements HttpClientFactory
@@ -20,7 +19,11 @@ class ClientFactory implements HttpClientFactory
 
     public function create(HttpClientConfig $config): Client
     {
-        if ($config->type === HttpClientConfig::TYPE_GUZZLE || $config->type === HttpClientConfig::TYPE_DEFAULT) {
+        if ($config->type === HttpClientConfig::TYPE_GUZZLE) {
+            if (! class_exists('\GuzzleHttp\Client')) {
+                throw new HttpClientTypeNotSupported('Guzzle Client unavailable');
+            }
+
             $wrapper = new ClientGuzzleWrapper(new GuzzleClient($config->params));
 
             return new MetricableClient($wrapper, $this->metrics);
